+++ test/unit/argparse_test.lua
@@ -166,7 +166,7 @@ g.test_sections = function()
     )
 end
 
-function g.test_priority()
+g.test_priority = function()
     utils.file_write(
         fio.pathjoin(g.tempdir, 'x.yml'),
         yaml.encode({
@@ -232,7 +232,7 @@ g.test_overrides = function()
     check('', {"TARANTOOL_A1=O1", "TARANTOOL_A1=O2"}, {a1 = 'O2', b2 = 2.2})
 end
 
-function g.test_appname()
+g.test_appname = function()
     utils.file_write(
         fio.pathjoin(g.tempdir, 'cfg.yml'),
         yaml.encode({
@@ -261,7 +261,7 @@ function g.test_appname()
     check('--app-name otherapp', {app_name = 'otherapp', y = '@otherapp'})
 end
 
-function g.test_confdir()
+g.test_confdir = function()
     local confd = fio.pathjoin(g.tempdir, 'conf.d')
     fio.mkdir(confd)
     utils.file_write(
@@ -333,7 +333,7 @@ g.test_badfile = function()
     t.assert_str_contains(ret.err, 'ParseConfigError: Missing section: app_name.harry')
 end
 
-function g.test_box_opts()
+g.test_box_opts = function()
     utils.file_write(
         fio.pathjoin(g.tempdir, 'cfg.yml'),
         yaml.encode({