+++ cartridge/argparse.lua
@@ -419,6 +419,10 @@ end
 --      arg3: true
 --    ...
 --
+-- Each option have a type: string, boolean, number.
+-- There is an ability to set multiple types for one option.
+-- Types are split by separator ``|``,  e.g. ``string|number``.
+--
 -- @function get_opts
 -- @tparam {argname=type,...} filter
 -- @treturn {argname=value,...}